+++ packages/bsky/src/api/app/bsky/graph/getFollows.ts
@@ -60,6 +60,7 @@ const skeleton = async (
     did: subjectDid,
     cursor: params.cursor,
     limit: params.limit,
+    sort: params.sort,
   })
   return {
     subjectDid,