+++ packages/internal/simple-store/src/index.ts
@@ -1,3 +1,4 @@
 export * from './cached-getter.js'
 export * from './simple-store.js'
+export * from './swallow-store-errors.js'
 export * from './util.js'