+++ packages/oauth/oauth-client-browser-example/src/providers/OAuthProvider.tsx
@@ -1,9 +1,11 @@
 import {
   type PropsWithChildren,
   createContext,
+  use,
   useCallback,
   useContext,
   useEffect,
+  useMemo,
   useState,
 } from 'react'
 import type { OAuthSession } from '@atproto/oauth-client-browser'
@@ -21,10 +23,10 @@ export type SignUpFunction = (
 export type SignOutFunction = () => Promise<void>
 
 export type OAuthValue = {
-  session: null | OAuthSession
-  isInitialized: boolean
-  isLoading: boolean
-  isSignedIn: boolean
+  /** State saved when we last left this app to complete the oauth flow */
+  state?: string | null
+  session?: OAuthSession
+
   signIn: SignInFunction
   signUp: SignUpFunction
   signOut: SignOutFunction
@@ -33,30 +35,13 @@ export type OAuthValue = {
 export const OAuthContext = createContext<null | OAuthValue>(null)
 OAuthContext.displayName = 'OAuthContext'
 
-export function OAuthProvider({ children }: PropsWithChildren) {
-  const [initialized, setInitialized] = useState(false)
-  const [loading, setLoading] = useState(false)
-  const [session, setSession] = useState<null | OAuthSession>(null)
+export type OAuthProviderProps = PropsWithChildren
 
-  useAbortableEffect(
-    (signal) => {
-      setInitialized(false)
-      setSession(null)
+export function OAuthProvider({ children }: OAuthProviderProps) {
+  const initSession = use(initPromise)
 
-      void initPromise
-        .then(async (result) => {
-          if (signal.aborted) return
-
-          if (result) setSession(result.session)
-        })
-        .finally(() => {
-          if (signal.aborted) return
-
-          setInitialized(true)
-        })
-    },
-    [initPromise],
-  )
+  const state = initSession?.state
+  const [session, setSession] = useState(initSession?.session)
 
   // Keep tabs in sync by listening to the oauth client's events and updating
   // the session state accordingly. The deletion part is needed because the
@@ -71,7 +56,7 @@ export function OAuthProvider({ children }: PropsWithChildren) {
         oauthEvents.addEventListener(
           'deleted',
           (evt) => {
-            if (evt.detail.sub === session.sub) setSession(null)
+            if (evt.detail.sub === session.sub) setSession(undefined)
           },
           { signal },
         )
@@ -113,67 +98,40 @@ export function OAuthProvider({ children }: PropsWithChildren) {
 
   const signIn = useCallback<SignInFunction>(
     async (input, options) => {
-      setLoading(true)
+      const session = await oauthClient
+        .restore(input, true)
+        .catch(async (_err) => oauthClient.signIn(input, options))
 
-      try {
-        const session = await oauthClient
-          .restore(input, true)
-          .catch(async (_err) => oauthClient.signIn(input, options))
-
-        setSession(session)
-      } finally {
-        setLoading(false)
-      }
+      setSession(session)
     },
     [oauthClient],
   )
 
   const signOut = useCallback<SignOutFunction>(async () => {
     if (session) {
-      setSession(null)
-      setLoading(true)
-      try {
-        await session.signOut()
-      } finally {
-        setLoading(false)
-      }
+      setSession(undefined)
+      await session.signOut()
     }
   }, [session])
 
   const signUp = useCallback<SignUpFunction>(
     async (input, options) => {
-      setLoading(true)
-      try {
-        const session = await oauthClient.signIn(input, {
-          ...options,
-          prompt: 'create',
-        })
+      const session = await oauthClient.signIn(input, {
+        ...options,
+        prompt: 'create',
+      })
 
-        setSession(session)
-      } finally {
-        setLoading(false)
-      }
+      setSession(session)
     },
     [oauthClient],
   )
 
-  return (
-    <OAuthContext.Provider
-      value={{
-        session,
-
-        isInitialized: initialized,
-        isLoading: !initialized || loading,
-        isSignedIn: !!session,
-
-        signIn,
-        signUp,
-        signOut,
-      }}
-    >
-      {children}
-    </OAuthContext.Provider>
+  const value = useMemo<OAuthValue | null>(
+    () => ({ session, state, signIn, signUp, signOut }),
+    [session, state, signIn, signUp, signOut],
   )
+
+  return <OAuthContext.Provider value={value}>{children}</OAuthContext.Provider>
 }
 
 export function useOAuthContext(hookName = useOAuthContext.name) {