+++ packages/ozone/src/api/report/getHistoricalStats.ts
@@ -1,3 +1,4 @@
+import { ForbiddenError } from '@atproto/xrpc-server'
 import type { AppContext } from '../../context.js'
 import type { Server } from '../../lexicon/index.js'
 import { viewHistoricalStats } from '../../report/views.js'
@@ -5,7 +6,7 @@ import { viewHistoricalStats } from '../../report/views.js'
 export default function (server: Server, ctx: AppContext) {
   server.tools.ozone.report.getHistoricalStats({
     auth: ctx.authVerifier.modOrAdminToken,
-    handler: async ({ params }) => {
+    handler: async ({ params, auth }) => {
       const {
         queueId,
         moderatorDid,
@@ -15,6 +16,10 @@ export default function (server: Server, ctx: AppContext) {
         limit,
         cursor,
       } = params
+
+      if (moderatorDid && !auth.credentials.isAdmin) {
+        throw new ForbiddenError('Unauthorized')
+      }
 
       const reportStatsService = ctx.reportStatsService(ctx.db)
       const result = await reportStatsService.getHistoricalStats({