+++ packages/bsync/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atproto/bsync
 
+## 0.0.36
+
+### Patch Changes
+
+- Updated dependencies [[`9e1da84`](https://github.com/bluesky-social/atproto/commit/9e1da84cc00a1cc5b453705539a2766403f3ee55)]:
+  - @atproto/syntax@0.7.2
+  - @atproto/common@0.7.2
+
 ## 0.0.35
 
 ### Patch Changes